Move Main Menu to its own file
